Task complete, but maybe need refactoring and migrate file type from .properties to another type.

+++ javasberschool/Lesson8/src/main/java/ru/anani/lesson8/annotation/Cache.java
@@ -11,6 +11,6 @@ public @interface Cache {
     CachedType cacheType();
     String fileNamePrefix() default "";
     boolean zip() default false;
-    long listList() default 16;
+    int listList() default 16;
     int[] ignoreIdentificationBy() default {};
 }
